move infinex config into its own file
