+++ src/ui-config/infinexConfig.ts
@@ -1,0 +1,14 @@
+import { InfinexConnectConfig, initInfinex } from '@infinex/connect-sdk';
+
+const infinexConfig: InfinexConnectConfig = {
+  appKey: 'aave',
+  debug: {
+    showLogs: true,
+    iframeOptions: {
+      shown: true,
+      width: '750',
+      height: '405',
+    },
+  },
+};
+export const infinexCore = initInfinex(infinexConfig);
